게시판 디테일 수정, 삭제, 취소 버튼 추가

+++ src/main/java/us/flower/dayary/controller/community/CommunityBoardController.java
@@ -100,10 +100,10 @@ public class CommunityBoardController {
 
 		// people name (사용자 정보)
 		Long peopleId = (Long) session.getAttribute("peopleId");
-		Long writerId = (Long) writer.getId();
+		Long writerId = writer.getId();
 
 		// 작성자 == 사용자 확인
-		if(peopleId==writerId){
+		if(peopleId.longValue()==writerId.longValue()){
 			model.addAttribute("writerFlag", TRUE);
 		}else{
 			model.addAttribute("writerFlag", FALSE);
